+++ services/protocol/programs/protocol/src/instructions/mod.rs
@@ -1,12 +1,12 @@
 pub mod create_meme_token;
 pub mod evolutions;
-pub mod trade_orca;
+pub mod trade_meteora;
 pub mod conviction_nfts;
 pub mod fee_distribution;
 
 // Re-export them for easy usage in lib.rs
 pub use create_meme_token::*;
 pub use evolutions::*;
-pub use trade_orca::*;
+pub use trade_meteora::*;
 pub use conviction_nfts::*;
 pub use fee_distribution::*;
